# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: manageCallAI API
  version: 0.1.0
  summary: Business-level telecom control plane API over FreeSWITCH
  description: |
    OpenAPI contract for the manageCallAI REST API.

    **MVP response envelope**: all implemented endpoints return `{ "data": ... }` for
    both single-resource and list responses. Pagination will be introduced later
    where required. Error responses return `{ "error": "<CODE>", "message": "<human text>",
    "request_id": "<id>" }` where `error` is a machine-readable RPC code (NOT_FOUND,
    INVALID_ARGUMENT, UNAUTHENTICATED, PERMISSION_DENIED, ALREADY_EXISTS,
    RESOURCE_EXHAUSTED, INTERNAL, UNAVAILABLE).

    **Auth model**:
    - Human-facing endpoints: `Authorization: Bearer <JWT>` (issued by `/auth/login`
      or `/auth/register`). JWT claims include `sub`, `tenant_id`, `email`, `role`.
    - Runtime/internal endpoints: `Authorization: Bearer <RUNTIME_API_TOKEN>` or
      header `x-managecallai-runtime-token: <token>`. These are called by
      FreeSWITCH and the Go ESL agent — not by end users.

    Only implemented endpoints are documented here. Future resources (outbound
    routes, prompts, CDRs, audit events, approvals, simulations) will be added
    when implemented.
servers:
  - url: https://api.example.com/api/v1
    description: Production placeholder
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /ivr-flows:
    get:
      tags:
        - IvrFlows
      summary: List IVR flows
      responses:
        "200":
          description: IVR flow list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrFlowListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags:
        - IvrFlows
      summary: Create IVR flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIvrFlowRequest"
      responses:
        "201":
          description: IVR flow created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrFlowDetailResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /ivr-flows/{flowId}:
    parameters:
      - $ref: "#/components/parameters/FlowIdParam"
    get:
      tags:
        - IvrFlows
      summary: Get IVR flow with versions
      responses:
        "200":
          description: IVR flow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrFlowDetailResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    patch:
      tags:
        - IvrFlows
      summary: Update IVR flow metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIvrFlowRequest"
      responses:
        "200":
          description: IVR flow updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrFlowResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /ivr-flows/{flowId}/versions:
    parameters:
      - $ref: "#/components/parameters/FlowIdParam"
    get:
      tags:
        - IvrFlows
      summary: List IVR flow versions
      responses:
        "200":
          description: Flow version list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowVersionListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags:
        - IvrFlows
      summary: Create draft version
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlowVersionDraftRequest"
      responses:
        "201":
          description: Draft version created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowVersionResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /ivr-flows/{flowId}/versions/{versionId}:
    parameters:
      - $ref: "#/components/parameters/FlowIdParam"
      - $ref: "#/components/parameters/VersionIdParam"
    get:
      tags:
        - IvrFlows
      summary: Get IVR flow version
      responses:
        "200":
          description: Flow version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowVersionResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    patch:
      tags:
        - IvrFlows
      summary: Update draft version definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlowVersionDraftRequest"
      responses:
        "200":
          description: Version definition updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowVersionResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /ivr-flows/{flowId}/versions/{versionId}/validate:
    parameters:
      - $ref: "#/components/parameters/FlowIdParam"
      - $ref: "#/components/parameters/VersionIdParam"
    post:
      tags:
        - IvrFlows
      summary: Validate flow version
      responses:
        "200":
          description: Validation passed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrFlowValidationResponse"
        "422":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrFlowValidationResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /ivr-flows/{flowId}/validate:
    parameters:
      - $ref: "#/components/parameters/FlowIdParam"
    post:
      tags:
        - IvrFlows
      summary: Validate current draft flow version
      responses:
        "200":
          description: Validation passed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrFlowValidationResponse"
        "422":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrFlowValidationResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /ivr-flows/{flowId}/simulate:
    parameters:
      - $ref: "#/components/parameters/FlowIdParam"
    post:
      tags:
        - IvrFlows
      summary: Simulate current draft flow version
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IvrFlowSimulationRequest"
      responses:
        "200":
          description: Simulation passed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrFlowSimulationResponse"
        "422":
          description: Simulation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrFlowSimulationResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /ivr-flows/{flowId}/versions/{versionId}/simulate:
    parameters:
      - $ref: "#/components/parameters/FlowIdParam"
      - $ref: "#/components/parameters/VersionIdParam"
    post:
      tags:
        - IvrFlows
      summary: Simulate specific flow version
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IvrFlowSimulationRequest"
      responses:
        "200":
          description: Simulation passed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrFlowSimulationResponse"
        "422":
          description: Simulation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrFlowSimulationResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /ivr-flows/{flowId}/versions/{versionId}/publish:
    parameters:
      - $ref: "#/components/parameters/FlowIdParam"
      - $ref: "#/components/parameters/VersionIdParam"
    post:
      tags:
        - IvrFlows
      summary: Publish validated or simulated flow version
      responses:
        "200":
          description: Flow published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrFlowPublishResponse"
        "202":
          description: Publish is pending approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrFlowPublishResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /ivr-flows/{flowId}/rollback:
    parameters:
      - $ref: "#/components/parameters/FlowIdParam"
    post:
      tags:
        - IvrFlows
      summary: Roll back to previous published version
      responses:
        "200":
          description: Flow rolled back
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrFlowPublishResponse"
        "202":
          description: Rollback is pending approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrFlowPublishResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /ivr-flows/{flowId}/history:
    parameters:
      - $ref: "#/components/parameters/FlowIdParam"
    get:
      tags:
        - IvrFlows
      summary: Get IVR flow validation, simulation, publish, and audit history
      responses:
        "200":
          description: IVR flow history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrFlowHistoryResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User JWT issued by /auth/login or /auth/register
    runtimeTokenBearer:
      type: http
      scheme: bearer
      description: RUNTIME_API_TOKEN as Bearer token (preferred)
    runtimeTokenHeader:
      type: apiKey
      in: header
      name: x-managecallai-runtime-token
      description: RUNTIME_API_TOKEN as custom header
  parameters:
...
    FlowIdParam:
      name: flowId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VersionIdParam:
      name: versionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          enum:
            - NOT_FOUND
            - INVALID_ARGUMENT
            - UNAUTHENTICATED
            - PERMISSION_DENIED
            - ALREADY_EXISTS
            - RESOURCE_EXHAUSTED
            - INTERNAL
            - UNAVAILABLE
        message:
          type: string
        request_id:
          type: string
      required:
        - error
        - message
        - request_id
...